+++ src/main/java/org/guzzing/studayserver/domain/calendar/service/AcademyCalendarService.java
@@ -2,6 +2,7 @@ package org.guzzing.studayserver.domain.calendar.service;
 
 import java.time.LocalDate;
 import java.util.List;
+
 import org.guzzing.studayserver.domain.calendar.model.AcademySchedule;
 import org.guzzing.studayserver.domain.calendar.model.AcademyTimeTemplate;
 import org.guzzing.studayserver.domain.calendar.repository.academyschedule.AcademyScheduleRepository;
@@ -20,7 +21,7 @@ import org.guzzing.studayserver.domain.calendar.service.dto.result.AcademyCalend
 import org.guzzing.studayserver.domain.calendar.service.dto.result.AcademyCalendarLoadToUpdateResult;
 import org.guzzing.studayserver.domain.calendar.service.dto.result.AcademyCalendarUpdateResults;
 import org.guzzing.studayserver.domain.dashboard.service.access.DashboardAccessService;
-import org.guzzing.studayserver.domain.dashboard.service.access.dto.DashboardScheduleAccessResult;
+import org.guzzing.studayserver.domain.dashboard.service.dto.response.DashboardScheduleAccessResult;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -30,17 +31,14 @@ public class AcademyCalendarService {
     private final AcademyScheduleRepository academyScheduleRepository;
     private final AcademyTimeTemplateRepository academyTimeTemplateRepository;
     private final PeriodicStrategy periodicStrategy;
-    private final DashboardAccessService dashboardAccessService;
 
     public AcademyCalendarService(
             AcademyScheduleRepository academyScheduleRepository,
             AcademyTimeTemplateRepository academyTimeTemplateRepository,
-            PeriodicStrategy periodicStrategy,
-            DashboardAccessService dashboardAccessService) {
+            PeriodicStrategy periodicStrategy) {
         this.academyScheduleRepository = academyScheduleRepository;
         this.academyTimeTemplateRepository = academyTimeTemplateRepository;
         this.periodicStrategy = periodicStrategy;
-        this.dashboardAccessService = dashboardAccessService;
     }
 
     @Transactional
@@ -125,10 +123,8 @@ public class AcademyCalendarService {
     public AcademyCalendarLoadToUpdateResult loadTimeTemplateToUpdate(Long academyScheduleId) {
         AcademyTimeTemplate academyTimeTemplate = academyScheduleRepository.findDistinctAcademyTimeTemplate(
                 academyScheduleId);
-        DashboardScheduleAccessResult dashboardScheduleAccessResult =
-                dashboardAccessService.getDashboardSchedule(academyTimeTemplate.getDashboardId());
 
-        return AcademyCalendarLoadToUpdateResult.of(academyTimeTemplate, dashboardScheduleAccessResult);
+        return AcademyCalendarLoadToUpdateResult.of(academyTimeTemplate);
     }
 
     @Transactional
